# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewPack API Documentation",
        "description": "Documentación completa de la API de NewPack",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get all customers",
                "operationId": "c680baeaff1a8944e0c9cbe5c6b5c1ac",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filter by customer name (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[last_name]",
                        "in": "query",
                        "description": "Filter by customer last name (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[email]",
                        "in": "query",
                        "description": "Filter by customer email (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[rfc]",
                        "in": "query",
                        "description": "Filter by customer RFC (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "description": "Filter by status (exact match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[client_type]",
                        "in": "query",
                        "description": "Filter by client type (exact match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[suburb_id]",
                        "in": "query",
                        "description": "Filter by suburb ID (exact match)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[search]",
                        "in": "query",
                        "description": "Search across multiple fields using the search scope",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customers list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "428d5533c8386f91d2baf4de69f01468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customers/{customer}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a customer by ID",
                "operationId": "dab8c06ba11e633bb424e2fca1f0fbac",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update a customer",
                "operationId": "f28e50b8a6a110ecab6f2aa45eca980b",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Soft delete a customer",
                "operationId": "eae002fbdbb9a6ef1275605e346d807f",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCustomerRequest": {
                "required": [
                    "name",
                    "last_name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Pérez García"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "5512345678",
                        "nullable": true
                    },
                    "phone_secondary": {
                        "type": "string",
                        "example": "5587654321",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "whatsapp": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "suburb_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "street": {
                        "type": "string",
                        "example": "Av. Insurgentes",
                        "nullable": true
                    },
                    "exterior_number": {
                        "type": "string",
                        "example": "123",
                        "nullable": true
                    },
                    "interior_number": {
                        "type": "string",
                        "example": "4B",
                        "nullable": true
                    },
                    "address_reference": {
                        "type": "string",
                        "example": "Entre calles X y Y",
                        "nullable": true
                    },
                    "tax_id": {
                        "description": "RFC - Tax ID",
                        "type": "string",
                        "example": "PEGJ800101AB1",
                        "nullable": true
                    },
                    "legal_name": {
                        "description": "Razón Social",
                        "type": "string",
                        "example": "Empresa SA de CV",
                        "nullable": true
                    },
                    "tax_system": {
                        "description": "Régimen Fiscal",
                        "type": "string",
                        "example": "601",
                        "nullable": true
                    },
                    "cfdi_use": {
                        "description": "Uso CFDI",
                        "type": "string",
                        "example": "G03",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended",
                            "blacklisted"
                        ],
                        "example": "active"
                    },
                    "client_type": {
                        "type": "string",
                        "enum": [
                            "individual",
                            "company"
                        ],
                        "example": "individual"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Cliente VIP",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateCustomerRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Pérez García"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "5512345678",
                        "nullable": true
                    },
                    "phone_secondary": {
                        "type": "string",
                        "example": "5587654321",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "whatsapp": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "suburb_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "street": {
                        "type": "string",
                        "example": "Av. Insurgentes",
                        "nullable": true
                    },
                    "exterior_number": {
                        "type": "string",
                        "example": "123",
                        "nullable": true
                    },
                    "interior_number": {
                        "type": "string",
                        "example": "4B",
                        "nullable": true
                    },
                    "address_reference": {
                        "type": "string",
                        "example": "Entre calles X y Y",
                        "nullable": true
                    },
                    "tax_id": {
                        "description": "RFC - Tax ID",
                        "type": "string",
                        "example": "PEGJ800101AB1",
                        "nullable": true
                    },
                    "legal_name": {
                        "description": "Razón Social",
                        "type": "string",
                        "example": "Empresa SA de CV",
                        "nullable": true
                    },
                    "tax_system": {
                        "description": "Régimen Fiscal",
                        "type": "string",
                        "example": "601",
                        "nullable": true
                    },
                    "cfdi_use": {
                        "description": "Uso CFDI",
                        "type": "string",
                        "example": "G03",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended",
                            "blacklisted"
                        ],
                        "example": "active"
                    },
                    "client_type": {
                        "type": "string",
                        "enum": [
                            "individual",
                            "company"
                        ],
                        "example": "individual"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Cliente VIP",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CustomerResource": {
                "title": "Customer",
                "description": "Customer resource with complete information",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "fullName": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com",
                        "nullable": true
                    },
                    "emailVerifiedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "5512345678",
                        "nullable": true
                    },
                    "phoneSecondary": {
                        "type": "string",
                        "example": "5587654321",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "whatsapp": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "suburbId": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "street": {
                        "type": "string",
                        "example": "Av. Insurgentes",
                        "nullable": true
                    },
                    "exteriorNumber": {
                        "type": "string",
                        "example": "123",
                        "nullable": true
                    },
                    "interiorNumber": {
                        "type": "string",
                        "example": "4B",
                        "nullable": true
                    },
                    "addressReference": {
                        "type": "string",
                        "example": "Entre calles X y Y",
                        "nullable": true
                    },
                    "fullAddress": {
                        "type": "string",
                        "example": "Av. Insurgentes 123, Int. 4B, Col. Centro, CDMX",
                        "nullable": true
                    },
                    "rfc": {
                        "type": "string",
                        "example": "PEPJ800101AB1",
                        "nullable": true
                    },
                    "razonSocial": {
                        "type": "string",
                        "example": "Empresa SA de CV",
                        "nullable": true
                    },
                    "regimenFiscal": {
                        "type": "string",
                        "example": "601",
                        "nullable": true
                    },
                    "usoCfdi": {
                        "type": "string",
                        "example": "G03",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended",
                            "blacklisted"
                        ],
                        "example": "active"
                    },
                    "clientType": {
                        "type": "string",
                        "enum": [
                            "individual",
                            "company"
                        ],
                        "example": "individual"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Cliente VIP",
                        "nullable": true
                    },
                    "createdBy": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "updatedBy": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "deletedBy": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "suburb": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SuburbResource"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "SuburbResource": {
                "title": "Suburb",
                "description": "Suburb resource",
                "properties": {
                    "id": {
                        "description": "ID del suburb",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del suburb",
                        "type": "string",
                        "example": "Centro"
                    },
                    "zipCodeId": {
                        "description": "ID del código postal asociado",
                        "type": "integer",
                        "example": 101
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingresa tu token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}